Second half of the application I guess

+++ rails/db/migrate/20211123015107_create_game_objects.rb
@@ -2,7 +2,7 @@ class CreateGameObjects < ActiveRecord::Migration[6.1]
   def change
     create_table :game_objects do |t|
       t.string :type
-      t.numeric :version
+      t.integer :version
       t.timestamps
     end
     create_table :entries do |t|
